feat(graph): remove double-tap-to-collapse on folder nodes
Double-tapping a folder body collapsed/expanded it, which fired too
easily by accident. Folder collapse remains available through the
deliberate affordances: the FolderHandleService chevron chip, the
vertical menu, and the folder-tree sidebar.

- Drop the `cy.on('dbltap', 'node[?isFolderNode]')` handler (and its now
  unused toggleFolderCollapse import) in setupBasicCytoscapeEventListeners.
- Retarget the folder-collapse e2e specs from synthetic dbltap emission to
  the surviving chevron affordance. Hoist `clickFolderChevron` (real on-
  screen click) into shared folder-test-helpers and add a programmatic
  `toggleFolderViaChevron` for mechanics-focused tests.
- Refresh the stale "dbltap-to-collapse" rationale on the AnchoredEditor
  folder-guard test; folder cy events stay enabled for hover/chip, not dbltap.
